Refactor widgets structure [WIP]
