--- src/actions/fetchApi.js
@@ -1,5 +1,0 @@
-import 'whatwg-fetch';
-
-const fetchApi = endpoint => fetch(`/api/${endpoint}`).then(response => response.json());
-
-export default fetchApi;
