+++ src/lib/fetchApi.js
@@ -1,0 +1,6 @@
+import 'whatwg-fetch';
+
+const fetchApi = endpoint =>
+  fetch(`/api/${endpoint}`).then(response => response.json());
+
+export default fetchApi;
