+++ src/widgets/NewRelicErrors/actions/newRelicErrors.js
@@ -1,19 +1,12 @@
 import 'whatwg-fetch';
 import { createAction } from 'redux-actions';
-import fetchApi from '../../../actions/fetchApi';
+import fetchApi from '../../../lib/fetchApi';
 
-export const requestNewRelicErrors = createAction('NEWRELICERRORS_REQUEST');
-export const receiveNewRelicErrors = createAction(
-  'NEWRELICERRORS_RECEIVE',
-  json => ({
-    previous: json.data.previous,
-    current: json.data.current,
-  }),
-);
+export const onReceive = createAction('NEWRELICERRORS_RECEIVE', json => ({
+  previous: json.data.previous,
+  current: json.data.current,
+}));
 
-export const fetchNewRelicErrors = () => dispatch => {
-  dispatch(requestNewRelicErrors());
-  return fetchApi('newRelicErrors').then(json =>
-    dispatch(receiveNewRelicErrors(json)),
-  );
+export const refresh = () => dispatch => {
+  fetchApi('newRelicErrors').then(json => dispatch(onReceive(json)));
 };
