+++ src/widgets/NewRelicErrors/actions/newRelicErrors.test.js
@@ -12,13 +12,13 @@ describe('a bugs difference receiving action', () => {
         current: 9,
       },
     };
-    expect(actions.receiveNewRelicErrors(json)).toEqual(expectedAction);
+    expect(actions.onReceive(json)).toEqual(expectedAction);
   });
 
   it('throws if parsing fails', () => {
     const json = { broken_data: { previous: 99, current: 9 } };
     expect(() => {
-      actions.receiveNewRelicErrors(json);
+      actions.onReceive(json);
     }).toThrow();
   });
 });
